LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/stretch/auto-margins-1-expected.txt

lorem
lorem

PASS [data-expected-width] 1
FAIL [data-expected-width] 2 assert_equals:
<div style="margin: 0 auto; min-width: stretch; width: 25px; background: lime;" data-expected-width="100" data-offset-x="0">lorem</div>
width expected 100 but got 25
